+++ main/window/window_xxx_background_menu.js
@@ -1,5 +1,5 @@
 ﻿'use strict';
-//28/07/26
+//10/09/26
 
 /* exported createBackgroundMenu */
 
@@ -110,13 +110,12 @@ function createBackgroundMenu(appendTo, parentMenu, options = { nameColors: fals
 					: folders.getBioArtistArtPath();
 				let input = option.newValue === 'path'
 					? Input.string('string', this.coverModeOptions.path, 'Enter TF expression or file path:\n\nMicrosoft MS-DOS wildcards are also allowed (at end of path), but only a single file will be matched. i.e.\n$directory_path(%PATH%)\\art\\*.jpg\n\nPaths starting with \'.\\profile\\\' are relative to foobar profile folder.' + (bLoadXXX ? '\nPaths starting with \'' + folders.xxxRootName + '\' are relative to this script\'s folder.' : '') + '\n\n\'%FB2K_PROFILE_PATH%\' or \'%PROFILE%\' may also be used.\n\nFor example:\n' + defTf + '\n\n\'DEFAULT\' applies default expression (above).', window.Name + ' (' + window.ScriptInfo.Name + '): Background file path', defTf)
-					: Input.string('string', this.coverModeOptions.path, 'Enter TF expression or folder path:\n\nMicrosoft MS-DOS wildcards are also allowed (at end of path), and all files matched by expression will be used. Note a folder path is equivalent to using [path]\\*.* i.e.\n$directory_path(%PATH%)\\art\\*.jpg\n\nPaths starting with \'.\\profile\\\' are relative to foobar profile folder.' + (bLoadXXX ? '\nPaths starting with \'' + folders.xxxRootName + '\' are relative to this script\'s folder.' : '') + '\n\n\'%FB2K_PROFILE_PATH%\' or \'%PROFILE%\' may also be used.\n\nFor example:\n' + defTf + '\n\n\'DEFAULT\' applies default expression (above).', window.Name + ' (' + window.ScriptInfo.Name + '): Background folder path', defTf);
+					: Input.string('string', this.coverModeOptions.path, 'Enter TF expression or folder path:\n\nMicrosoft MS-DOS wildcards are also allowed (at end of path), and all files matched by expression will be used. Note a folder path is equivalent to using [path]\\*.* i.e.\n$directory_path(%PATH%)\\art\\*.jpg\n\nPaths starting with \'.\\profile\\\' are relative to foobar profile folder.' + (bLoadXXX ? '\nPaths starting with \'' + folders.xxxRootName + '\' are relative to this script\'s folder.' : '') + '\n\n\'%FB2K_PROFILE_PATH%\' or \'%PROFILE%\' may also be used.\n\nFor example:\n' + defTf + '\n\n\'DEFAULT\' applies default expression (above).', window.Name + ' (' + window.ScriptInfo.Name + '): Background folder path', defTf, void(0), void(0), defTf);
 				if (input === null) {
 					if (this.coverMode === option.newValue) { return false; }
 					else if (Input.isLastEqual) { input = Input.lastInput; }
 					else { return false; }
 				}
-				input = input.replace('DEFAULT', defTf);
 				this.changeConfig({ config: { coverModeOptions: { path: input } }, callbackArgs: { bSaveProperties: true } });
 			}
 			return true;